integrate dev_mixer changes in qq & power to bivar_mixer
